book/fsl-example/music_fsl/requirements.txt
# data
mirdata
librosa

# training
torch
numpy
torchaudio
torchmetrics
pytorch-lightning
tensorboard
argbind

# display
torchvision
sklearn
umap-learn
pandas
plotly
kaleido
